fix(lastfm): service worker
